Fixes; Added Manjaro

+++ dl/setup.sh
@@ -11,23 +11,28 @@ get_config_repo() {
 	echo "Downloading and installing dotfiles..."
 	git clone https://github.com/williki/configs.git
 	mkdir -p ~/.config/htop
-	mv -i configs/htoprc ~/.config/htop/htoprc
+	mv -i configs/.config/htop/htoprc ~/.config/htop/htoprc
 	mv -i configs/.zshrc ~
 	mv -i configs/.vimrc ~
 	rm -rf configs
 }
 
 install_core_progs() {
-	
+
 	# Arch based OS
 	if [ -f /usr/bin/pacman ]; then
 		echo "Arch based OS detected, installing core progs..."
-		sudo pacman -Syu --noconfirm binutils curl git htop vim zsh zsh-autosuggestions zsh-syntax-highlightning base-devel
-		git clone https://aur.archlinux.org/yay.git
-		cd yay
-		yes | makepkg -si
-		cd ..
-		rm -rf yay
+		sudo pacman -Syu --noconfirm binutils curl git htop vim zsh zsh-autosuggestions zsh-syntax-highlighting base-devel
+		# On Manjaro yay is in com,unity repo
+		if [ -f /.manjaro-tools ]; then
+			sudo pacman -S yay
+		else
+			git clone https://aur.archlinux.org/yay.git
+			cd yay
+			yes | makepkg -si
+			cd ..
+			rm -rf yay
+		fi
 	fi
 
 	# Debian based OS
